backend/requirements.txt
flask
flask_cors
langchain_google_genai
langchain_together
langchain
python-dotenv